refactor: moved registering field in field group to `useField`
can be used with component and composition API now
